criando Usuário é melhorando api com DTO

+++ devdojo-essentials-02/src/main/java/br/com/devdojo/essentials/repository/AnimeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AnimeRepository extends JpaRepository<Anime,Long> {
+public interface AnimeRepository extends JpaRepository<Anime,Integer> {
 
     List<Anime> findByName(String name);
